+++ include/franka/rate_limiting.h
@@ -123,11 +123,14 @@ constexpr double kMinElbowVelocity = -kMaxElbowVelocity;
  *
  * @return Upper limits of joint velocity at the given joint position.
  *
- * @deprecated This function is deprecated and will be removed in future versions. 
- * Use `Robot::getUpperJointVelocityLimits(const std::array<double, 7UL> &joint_positions)` instead (if applicable).
+ * @deprecated This function is deprecated and will be removed in future versions.
+ * Use `Robot::getUpperJointVelocityLimits(const std::array<double, 7UL> &joint_positions)` instead
+ * (if applicable).
  */
-[[deprecated("Use Robot::getUpperJointVelocityLimits(const std::array<double, 7UL> &joint_positions) instead.")]]
-inline std::array<double, 7> computeUpperLimitsJointVelocity(
+[[deprecated(
+    "Use Robot::getUpperJointVelocityLimits(const std::array<double, 7UL> &joint_positions) "
+    "instead with system images >= 5.9.0.")]] inline std::array<double, 7>
+computeUpperLimitsJointVelocity(
     const std::array<double, 7>& q) {  // NOLINT(readability-identifier-length)
   return std::array<double, 7>{
       std::min(2.62, std::max(0.0, -0.30 + std::sqrt(std::max(0.0, 12.0 * (2.75010 - q[0]))))) -
@@ -157,11 +160,14 @@ inline std::array<double, 7> computeUpperLimitsJointVelocity(
  *
  * @return Lower limits of joint velocity at the given joint position.
  *
- * @deprecated This function is deprecated and will be removed in future versions. 
- * Use `Robot::getlowerJointVelocityLimits(const std::array<double, 7UL> &joint_positions)` instead (if applicable).
+ * @deprecated This function is deprecated and will be removed in future versions.
+ * Use `Robot::getlowerJointVelocityLimits(const std::array<double, 7UL> &joint_positions)` instead
+ * (if applicable).
  */
-[[deprecated("Use Robot::getlowerJointVelocityLimits(const std::array<double, 7UL> &joint_positions) instead.")]]
-inline std::array<double, 7> computeLowerLimitsJointVelocity(
+[[deprecated(
+    "Use Robot::getlowerJointVelocityLimits(const std::array<double, 7UL> &joint_positions) "
+    "instead with system images >= 5.9.0.")]] inline std::array<double, 7>
+computeLowerLimitsJointVelocity(
     const std::array<double, 7>& q) {  // NOLINT(readability-identifier-length)
   return std::array<double, 7>{
       std::max(-2.62, std::min(0.0, 0.30 - std::sqrt(std::max(0.0, 12.0 * (2.750100 + q[0]))))) +